plug-ins: Handle little-endian PSD data correctly
Fix parsing of PSD files stored using IBM PC (little-endian) byte
order. The previous implementation incorrectly read image data, group
information, and other properties when loading little-endian files.

Refactor PSD reading code to use endian-aware wrapper functions for the
various data types, providing consistent handling of both little-endian
and big-endian PSD files and reducing duplication in the parser.

+++ plug-ins/file-psd/psd.h
@@ -617,8 +617,8 @@ typedef struct
 {
   guint32    size;
   guint32    ver;
-  guint32    blur;
-  guint32    intensity;
+  gfloat     blur;
+  gfloat     intensity;
   guint16    color[5];
   gchar      blendsig[4];
   guint32    effect;
